*** CONVERT PCB'S BASE UNITS TO NANOMETERS ***
Convert base units to nm, change Coord from int to long,
change LARGE_VALUE from a magic number to (LONG_MAX / 2 - 1).

Fixes-bug: lp-772027

+++ src/global.h
@@ -64,7 +64,7 @@ typedef struct AttributeListType AttributeListType, *AttributeListTypePtr;
 typedef struct unit Unit;
 typedef struct increments Increments;
 
-typedef int Coord;		/* pcb base unit */
+typedef long Coord;		/* pcb base unit */
 typedef double Angle;		/* degrees */
 
 #include "hid.h"
